requirements.txt
langchain
pickle5
streamlit_chat
streamlit
python-dotenv